# docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: IoT Monitoring Platform API
  version: "1.0.0-draft"
  description: >
    Contrato de referencia de la Etapa 7 (API_DESIGN.md). Cubre el alcance del MVP
    (Etapa 1/PRODUCT_REQUIREMENTS.md sección 9.1). No incluye comandos remotos,
    firmware, API keys ni informes (V2/Futuro). Los recursos de Directorio IoT bajo
    /platform se muestran solo con `gateways` como representativo; el resto de
    recursos (installations, zones, devices, sensors, channels) se mecaniza igual
    (ver API_DESIGN.md sección 2).
servers:
  - url: https://api.example.com/v1
    description: Producción (dominio real a definir en Etapa 9)
  - url: https://staging-api.example.com/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /installations:
    get:
      tags: [installations]
      summary: Listar instalaciones (con alcance aplicado, PERMISSIONS.md sección 2)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstallationPage" }
    post:
      tags: [installations]
      summary: Crear instalación
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InstallationCreate" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Installation" }
  /installations/{installationId}:
    get:
      tags: [installations]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Installation" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [installations]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InstallationCreate" }
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Installation" }
    delete:
      tags: [installations]
      summary: Baja lógica
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "204": { description: Sin contenido }
  /installations/{installationId}/zones:
    get:
      tags: [zones]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Zone" }
    post:
      tags: [zones]
      parameters:
        - $ref: "#/components/parameters/PathId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ZoneCreate" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Zone" }
...
  /zones/{zoneId}:
    get:
      tags: [zones]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Zone" }
    patch:
      tags: [zones]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ZoneCreate" }
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Zone" }
    delete:
      tags: [zones]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "204": { description: Sin contenido }
  /gateways:
    get:
      tags: [gateways]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GatewayPage" }
    post:
      tags: [gateways]
      summary: Pre-registrar un gateway (concentrador LoRa o estación directa, ADR-0004)
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GatewayCreate" }
      responses:
        "201":
          description: >
            Creado. El campo `credentialSecret` solo aparece en esta respuesta
            (API_DESIGN.md sección 7) — no se puede recuperar después.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GatewayCreated" }
  /gateways/{gatewayId}:
    get:
      tags: [gateways]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Gateway" }
    patch:
      tags: [gateways]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                heartbeatIntervalSeconds: { type: integer, nullable: true }
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Gateway" }
    delete:
      tags: [gateways]
      summary: Deshabilitar gateway
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "204": { description: Sin contenido }
  /gateways/{gatewayId}/rotate-credential:
    post:
      tags: [gateways]
      summary: Revocar la credencial activa y emitir una nueva
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200":
          description: "Éxito. Igual que en la creación, el secreto solo se muestra aquí."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GatewayCreated" }
  /gateways/{gatewayId}/devices:
    get:
      tags: [devices]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Device" }
    post:
      tags: [devices]
      parameters:
        - $ref: "#/components/parameters/PathId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeviceCreate" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
  /devices/{deviceId}:
    get:
      tags: [devices]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
    patch:
      tags: [devices]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                zoneId: { type: string, format: uuid }
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
    delete:
      tags: [devices]
      summary: Deshabilitar dispositivo
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "204": { description: Sin contenido }
  /devices/{deviceId}/sensors:
    get:
      tags: [sensors]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Sensor" }
    post:
      tags: [sensors]
      summary: Registrar un sensor (máximo 4 por dispositivo, FUNCTIONAL_REQUIREMENTS.md sección 7)
      parameters:
        - $ref: "#/components/parameters/PathId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SensorCreate" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sensor" }
  /sensors/{sensorId}:
    get:
      tags: [sensors]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sensor" }
    delete:
      tags: [sensors]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "204": { description: Sin contenido }
  /sensors/{sensorId}/channels:
    get:
      tags: [channels]
      summary: Canales del sensor (creados automáticamente por ingestión, MQTT_PROTOCOL.md sección 9)
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Channel" }
  /channels/{channelId}:
    patch:
      tags: [channels]
      summary: Sobrescribir el umbral de alerta de este canal (override del valor por defecto de la organización)
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                alertThresholdMin: { type: number, nullable: true }
                alertThresholdMax: { type: number, nullable: true }
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Channel" }
...
  /platform/organizations/{organizationId}/gateways:
    get:
      tags: [platform]
      summary: >
        Directorio IoT global (ADR-0005) — representativo de instalaciones/zonas/
        dispositivos/sensores/canales, que se mecanizan igual (API_DESIGN.md sección 16).
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200":
          description: Éxito
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Gateway" }
    post:
      tags: [platform]
      parameters:
        - $ref: "#/components/parameters/PathId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GatewayCreate" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GatewayCreated" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: Identificador de la ruta (nombre real varía por endpoint, p. ej. installationId, gatewayId)
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: API_DESIGN.md sección 6
  responses:
...
    Forbidden:
      description: Autenticado pero sin permiso (PERMISSIONS.md)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, example: "validation-error" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
...
    Installation:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        locationText: { type: string, nullable: true }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        status: { type: string, enum: [active, inactive] }
    InstallationCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        locationText: { type: string, nullable: true }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
    InstallationPage:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Installation" } }
        meta: { $ref: "#/components/schemas/PageMeta" }
    Zone:
      type: object
      properties:
        id: { type: string, format: uuid }
        installationId: { type: string, format: uuid }
        name: { type: string }
        zoneType: { type: string, nullable: true }
    ZoneCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        zoneType: { type: string, nullable: true }
    Gateway:
      type: object
      properties:
        id: { type: string, format: uuid }
        installationId: { type: string, format: uuid }
        externalIdentifier: { type: string }
        name: { type: string }
        connectivityType: { type: string, enum: [lora_concentrator, direct_nbiot, direct_other] }
        status: { type: string, enum: [not_provisioned, online, offline, disabled] }
        lastSeenAt: { type: string, format: date-time, nullable: true }
        heartbeatIntervalSeconds: { type: integer, nullable: true }
    GatewayCreate:
      type: object
      required: [installationId, name, connectivityType]
      properties:
        installationId: { type: string, format: uuid }
        name: { type: string }
        connectivityType: { type: string, enum: [lora_concentrator, direct_nbiot, direct_other] }
    GatewayCreated:
      allOf:
        - $ref: "#/components/schemas/Gateway"
        - type: object
          properties:
            credentialUsername: { type: string }
            credentialSecret:
              type: string
              description: Solo se devuelve en esta respuesta (API_DESIGN.md sección 7)
    GatewayPage:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Gateway" } }
        meta: { $ref: "#/components/schemas/PageMeta" }
    Device:
      type: object
      properties:
        id: { type: string, format: uuid }
        gatewayId: { type: string, format: uuid }
        zoneId: { type: string, format: uuid }
        externalIdentifier: { type: string }
        name: { type: string }
        status: { type: string, enum: [not_provisioned, online, offline, disabled] }
        lastSeenAt: { type: string, format: date-time, nullable: true }
    DeviceCreate:
      type: object
      required: [zoneId, externalIdentifier, name]
      properties:
        zoneId: { type: string, format: uuid }
        externalIdentifier: { type: string }
        name: { type: string }
    Sensor:
      type: object
      properties:
        id: { type: string, format: uuid }
        deviceId: { type: string, format: uuid }
        externalIdentifier: { type: string }
        label: { type: string, nullable: true }
    SensorCreate:
      type: object
      required: [externalIdentifier]
      properties:
        externalIdentifier: { type: string }
        label: { type: string, nullable: true }
    Channel:
      type: object
      properties:
        id: { type: string, format: uuid }
        sensorId: { type: string, format: uuid }
        channelTypeCode: { type: string }
        alertThresholdMin: { type: number, nullable: true }
        alertThresholdMax: { type: number, nullable: true }
...
    PageMeta:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...